contrib dir: removed unessesary old forms of Integer type
Changed boxing/autoboxing with Integer()/valueOf()/intValue() to raw int values or parseInt(some string)

+++ contrib/j2ee.weblogic9/test/unit/src/org/netbeans/modules/j2ee/weblogic9/WLPluginPropertiesTest.java
@@ -57,8 +57,8 @@ public class WLPluginPropertiesTest extends NbTestCase {
         createJar(file, "Implementation-Version: 10.0.0.1");
         Version version = WLPluginProperties.getServerVersion(baseFolder);
         assertEquals("10.0.0.1", version.toString());
-        assertEquals(10, version.getMajor().intValue());
-        assertEquals(1, version.getUpdate().intValue());
+        assertEquals(10, version.getMajor());
+        assertEquals(1, version.getUpdate());
         assertTrue(file.delete());
     }
